backend/requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.3.1
Pillow>=10.0.0
python-dotenv==0.19.0
Werkzeug==2.0.1
pymongo==4.3.3
dnspython==2.3.0
certifi>=2023.7.22
pyOpenSSL>=23.2.0